+++ src/components/Modal/index.tsx
@@ -2,8 +2,7 @@ import React, {memo, ReactElement, useCallback, useEffect, useRef} from 'react';
 import {Animated, StyleSheet, View} from 'react-native';
 import {ScrollView, TouchableWithoutFeedback} from '../../conversions';
 import {useColor, useDriver, useDropShadow} from '../../hooks';
-import {useRootSelector} from '../../utils';
-import {SoundManager} from '../../utils/Sound';
+import {SoundManager, useRootSelector} from '../../utils';
 import {Card} from '../Card';
 
 type ModalProps = {